backend/pdf/requirements.txt
psycopg2-binary
python-dotenv
reportlab
